@@ -38,99 +38,68 @@
    knowledge of the CeCILL license and that you accept its terms.
  */
 
+#include <Bpp/Seq/Alphabet/AlphabetTools.h>
 #include "DataSet.h"
 
 using namespace bpp;
 using namespace std;
 
-// ** Class constructor: *******************************************************/
-
-DataSet::DataSet() : analyzedLoci_(0),
-  analyzedSequences_(0),
-  localities_(vector<Locality<double>*>()),
-  groups_(vector<Group*>()) {}
-
 /******************************************************************************/
 
-DataSet::DataSet(const DataSet& ds) : analyzedLoci_(0),
-  analyzedSequences_(0),
-  localities_(vector<Locality<double>*>()),
-  groups_(vector<Group*>())
+DataSet::DataSet(const DataSet& ds) :
+  analyzedLoci_(nullptr),
+  sequenceAlphabet_(ds.sequenceAlphabet_),
+  localities_(),
+  groups_()
 {
-  if (ds.analyzedLoci_ != 0)
-    analyzedLoci_ = new AnalyzedLoci(*(ds.analyzedLoci_));
-  if (ds.analyzedSequences_ != 0)
-    analyzedSequences_ = new AnalyzedSequences(*(ds.analyzedSequences_));
-  if (ds.localities_.size() != 0)
-    for (size_t i = 0; i < ds.localities_.size(); i++)
-    {
-      localities_.push_back(new Locality<double>(*(ds.localities_[i])));
-    }
-  if (ds.groups_.size() != 0)
-    for (size_t i = 0; i < ds.groups_.size(); i++)
-    {
-      groups_.push_back(new Group(*(ds.groups_[i])));
-    }
+  if (ds.analyzedLoci_)
+    analyzedLoci_.reset(ds.analyzedLoci_->clone());
+  for (const auto& locality : ds.localities_)
+    localities_.push_back(unique_ptr<Locality<double>>(locality->clone()));
+  for (const auto& group : ds.groups_)
+    groups_.push_back(unique_ptr<Group>(group->clone()));
 }
 
 /******************************************************************************/
 
 DataSet& DataSet::operator=(const DataSet& ds)
 {
-  if (ds.analyzedLoci_ != 0)
-    analyzedLoci_ = new AnalyzedLoci(*(ds.analyzedLoci_));
-  if (ds.analyzedSequences_ != 0)
-    analyzedSequences_ = new AnalyzedSequences(*(ds.analyzedSequences_));
-  if (ds.localities_.size() != 0)
-    for (size_t i = 0; i < ds.localities_.size(); i++)
-    {
-      localities_.push_back(new Locality<double>(*(ds.localities_[i])));
-    }
-  if (ds.groups_.size() != 0)
-    for (size_t i = 0; i < ds.groups_.size(); i++)
-    {
-      groups_.push_back(new Group(*(ds.groups_[i])));
-    }
+  if (ds.analyzedLoci_)
+    analyzedLoci_.reset(ds.analyzedLoci_->clone());
+  else
+    analyzedLoci_.reset(nullptr);
+
+  sequenceAlphabet_ = ds.sequenceAlphabet_;
+ 
+  localities_.clear(); 
+  for (const auto& locality : ds.localities_)
+    localities_.push_back(unique_ptr<Locality<double>>(locality->clone()));
+
+  groups_.clear();
+  for (const auto& group : ds.groups_)
+    groups_.push_back(unique_ptr<Group>(group->clone()));
+  
   return *this;
 }
 
-// ** Class destructor: *******************************************************/
-DataSet::~DataSet()
-{
-  if (getNumberOfGroups() > 0)
-    for (size_t i = 0; i < getNumberOfGroups(); i++)
-    {
-      delete groups_[i];
-    }
-  if (analyzedLoci_ != 0)
-    delete analyzedLoci_;
-  if (getNumberOfLocalities() > 0)
-    for (size_t i = 0; i < getNumberOfLocalities(); i++)
-    {
-      delete localities_[i];
-    }
-  if (analyzedSequences_ != 0)
-    delete analyzedSequences_;
-}
-
-// ** Other methodes: *********************************************************/
+// ** Other methods: *********************************************************/
 
 // Dealing with Localities ---------------------------------
-void DataSet::addLocality(Locality<double>& locality)
+void DataSet::addLocality(const Locality<double>& locality)
 {
-  for (size_t i = 0; i < localities_.size(); i++)
+  for (const auto& existingLocality : localities_)
   {
-    if (localities_[i]->getName() == locality.getName())
+    if (existingLocality->getName() == locality.getName())
       throw BadIdentifierException("DataSet::addLocality: locality name already in use.", locality.getName());
   }
-  localities_.push_back(new Locality<double>(locality));
+  localities_.push_back(make_unique<Locality<double>>(locality));
 }
 
 /******************************************************************************/
 
 size_t DataSet::getLocalityPosition(const std::string& name) const
 {
-  for (size_t i = 0; i < localities_.size(); i++)
+  for (size_t i = 0; i < localities_.size(); ++i)
   {
     if (localities_[i]->getName() == name)
       return i;
@@ -140,16 +109,25 @@ size_t DataSet::getLocalityPosition(const std::string& name) const
 
 /******************************************************************************/
 
-const Locality<double>& DataSet::getLocalityAtPosition(size_t locality_position) const
+shared_ptr<const Locality<double>> DataSet::getLocalityAtPosition(size_t localityPosition) const
 {
-  if (locality_position >= localities_.size())
-    throw IndexOutOfBoundsException("DataSet::getLocalityAtPosition: locality_position out of bounds.", locality_position, 0, localities_.size());
-  return *(localities_[locality_position]);
+  if (localityPosition >= localities_.size())
+    throw IndexOutOfBoundsException("DataSet::getLocalityAtPosition: localityPosition out of bounds.", localityPosition, 0, localities_.size());
+  return localities_[localityPosition];
 }
 
 /******************************************************************************/
 
-const Locality<double>& DataSet::getLocalityByName(const std::string& name) const
+const Locality<double>& DataSet::localityAtPosition(size_t localityPosition) const
+{
+  if (localityPosition >= localities_.size())
+    throw IndexOutOfBoundsException("DataSet::getLocalityAtPosition: localityPosition out of bounds.", localityPosition, 0, localities_.size());
+  return *localities_[localityPosition];
+}
+
+/******************************************************************************/
+
+shared_ptr<const Locality<double>> DataSet::getLocalityByName(const std::string& name) const
 {
   try
   {
@@ -163,12 +141,25 @@ const Locality<double>& DataSet::getLocalityByName(const std::string& name) cons
 
 /******************************************************************************/
 
-void DataSet::deleteLocalityAtPosition(size_t locality_position)
+const Locality<double>& DataSet::localityByName(const std::string& name) const
 {
-  if (locality_position >= localities_.size())
-    throw IndexOutOfBoundsException("DataSet::deleteLocalityAtPosition: locality_position out of bounds.", locality_position, 0, localities_.size());
-  delete localities_[locality_position];
-  localities_.erase(localities_.begin() + static_cast<ptrdiff_t>(locality_position));
+  try
+  {
+    return localityAtPosition(getLocalityPosition(name));
+  }
+  catch (LocalityNotFoundException& lnfe)
+  {
+    throw LocalityNotFoundException("DataSet::getLocalityByName: Locality not found.", name);
+  }
+}
+
+/******************************************************************************/
+
+void DataSet::deleteLocalityAtPosition(size_t localityPosition)
+{
+  if (localityPosition >= localities_.size())
+    throw IndexOutOfBoundsException("DataSet::deleteLocalityAtPosition: localityPosition out of bounds.", localityPosition, 0, localities_.size());
+  localities_.erase(localities_.begin() + static_cast<ptrdiff_t>(localityPosition));
 }
 
 /******************************************************************************/
@@ -187,112 +178,98 @@ void DataSet::deleteLocalityByName(const std::string& name)
 
 /******************************************************************************/
 
-size_t DataSet::getNumberOfLocalities() const
-{
-  return localities_.size();
-}
-
-/******************************************************************************/
-
-bool DataSet::hasLocality() const
-{
-  return getNumberOfLocalities() > 0;
-}
-
-/******************************************************************************/
-
 // Dealing with groups -------------------------------------
 void DataSet::addGroup(const Group& group)
 {
-  for (size_t i = 0; i < groups_.size(); i++)
+  for (const auto& existingGroup : groups_)
   {
-    if (group.getGroupId() == groups_[i]->getGroupId())
+    if (group.getGroupId() == existingGroup->getGroupId())
       throw BadIdentifierException("DataSet::addGroup: group id already in use.", group.getGroupId());
   }
-  groups_.push_back(new Group(group));
+  groups_.push_back(make_unique<Group>(group));
 }
 
 /******************************************************************************/
 
-void DataSet::addEmptyGroup(size_t group_id)
+void DataSet::addEmptyGroup(size_t groupId)
+{
+  for (const auto& existingGroup : groups_)
+  {
+    if (groupId == existingGroup->getGroupId())
+      throw BadIdentifierException("DataSet::addEmptyGroup: groupId already in use.", groupId);
+  }
+  groups_.push_back(make_unique<Group>(groupId));
+}
+
+/******************************************************************************/
+
+const Group& DataSet::getGroupById(size_t groupId) const
 {
   for (size_t i = 0; i < groups_.size(); i++)
+  for (const auto& group : groups_)
   {
-    if (group_id == groups_[i]->getGroupId())
-      throw BadIdentifierException("DataSet::addEmptyGroup: group_id already in use.", group_id);
+    if (groupId == group->getGroupId())
+      return *group;
   }
-  groups_.push_back(new Group(group_id));
+  throw GroupNotFoundException("DataSet::getGroupById: groupId not found.", groupId);
 }
 
 /******************************************************************************/
 
-const Group& DataSet::getGroupById(size_t group_id) const
-{
-  for (size_t i = 0; i < groups_.size(); i++)
-  {
-    if (group_id == groups_[i]->getGroupId())
-      return *(groups_[i]);
-  }
-  throw GroupNotFoundException("DataSet::getGroupById: group_id not found.", group_id);
-}
-
-/******************************************************************************/
-
-string DataSet::getGroupName(size_t group_id) const
+string DataSet::getGroupName(size_t groupId) const
 {
   string name;
-  name = getGroupById(group_id).getGroupName();
+  name = getGroupById(groupId).getGroupName();
   if (!name.empty() )
     return name;
   else
-    return TextTools::toString(group_id);
-  throw GroupNotFoundException("DataSet::getGroupName: group_id not found.", group_id);
+    return TextTools::toString(groupId);
+  throw GroupNotFoundException("DataSet::getGroupName: groupId not found.", groupId);
 }
 
 /******************************************************************************/
 
-void DataSet::setGroupName(size_t group_id, const std::string& group_name) const
+void DataSet::setGroupName(size_t groupId, const std::string& group_name) const
 {
   for (size_t i = 0; i < groups_.size(); i++)
   {
-    if (group_id == groups_[i]->getGroupId())
+    if (groupId == groups_[i]->getGroupId())
     {
       groups_[i]->setGroupName(group_name);
       return;
     }
   }
-  throw GroupNotFoundException("DataSet::setGroupName: group_id not found.", group_id);
+  throw GroupNotFoundException("DataSet::setGroupName: groupId not found.", groupId);
 }
 
 /******************************************************************************/
 
-size_t DataSet::getGroupPosition(size_t group_id) const
+size_t DataSet::getGroupPosition(size_t groupId) const
 {
   for (size_t i = 0; i < groups_.size(); i++)
   {
-    if (group_id == groups_[i]->getGroupId())
+    if (groupId == groups_[i]->getGroupId())
       return i;
   }
-  throw GroupNotFoundException("DataSet::getGroupPosition: group_id not found.", group_id);
+  throw GroupNotFoundException("DataSet::getGroupPosition: groupId not found.", groupId);
 }
 
 /******************************************************************************/
 
-const Group& DataSet::getGroupAtPosition(size_t group_position) const
+const Group& DataSet::getGroupAtPosition(size_t groupPosition) const
 {
-  if (group_position >= groups_.size())
-    throw IndexOutOfBoundsException("DataSet::getGroup.", group_position, 0, groups_.size());
-  return *(groups_[group_position]);
+  if (groupPosition >= groups_.size())
+    throw IndexOutOfBoundsException("DataSet::getGroup.", groupPosition, 0, groups_.size());
+  return *(groups_[groupPosition]);
 }
 
 /******************************************************************************/
 
-void DataSet::deleteGroupAtPosition(size_t group_position)
+void DataSet::deleteGroupAtPosition(size_t groupPosition)
 {
-  if (group_position >= groups_.size())
-    throw IndexOutOfBoundsException("DataSet::deleteGroup.", group_position, 0, groups_.size());
-  delete groups_[group_position];
-  groups_.erase(groups_.begin() + static_cast<ptrdiff_t>(group_position));
+  if (groupPosition >= groups_.size())
+    throw IndexOutOfBoundsException("DataSet::deleteGroup.", groupPosition, 0, groups_.size());
+  groups_.erase(groups_.begin() + static_cast<ptrdiff_t>(groupPosition));
 }
 
 /******************************************************************************/
@@ -335,27 +312,27 @@ void DataSet::mergeTwoGroups(size_t source_id, size_t target_id)
 
 /******************************************************************************/
 
-void DataSet::mergeGroups(std::vector<size_t>& group_ids)
+void DataSet::mergeGroups(std::vector<size_t>& groupIds)
 {
   // Test if all group id exists in the DataSet
-  for (size_t i = 0; i < group_ids.size(); i++)
+  for (size_t i = 0; i < groupIds.size(); i++)
   {
     try
     {
-      getGroupById(group_ids[i]);
+      getGroupById(groupIds[i]);
     }
     catch (GroupNotFoundException& e)
     {
-      throw GroupNotFoundException("DataSet::mergeGroups: group not found.", group_ids[i]);
+      throw GroupNotFoundException("DataSet::mergeGroups: group not found.", groupIds[i]);
     }
   }
   // Sort the group id
-  sort(group_ids.begin(), group_ids.end());
+  sort(groupIds.begin(), groupIds.end());
   // Merge all the groups in the first
-  size_t pos_first = getGroupPosition(group_ids[0]);
-  for (size_t i = 1; i < group_ids.size(); i++)
+  size_t pos_first = getGroupPosition(groupIds[0]);
+  for (size_t i = 1; i < groupIds.size(); i++)
   {
-    size_t pos_current = getGroupPosition(group_ids[i]);
+    size_t pos_current = getGroupPosition(groupIds[i]);
     for (size_t j = 0; j < getGroupAtPosition(pos_current).getNumberOfIndividuals(); j++)
     {
       groups_[pos_first]->addIndividual(getGroupAtPosition(pos_current).getIndividualAtPosition(j));
@@ -366,36 +343,36 @@ void DataSet::mergeGroups(std::vector<size_t>& group_ids)
 
 /******************************************************************************/
 
-void DataSet::splitGroup(size_t group_id, std::vector<size_t> individuals_selection)
+void DataSet::splitGroup(size_t groupId, vector<size_t> individualSelection)
 {
-  size_t source_pos;
+  size_t sourcePos;
   try
   {
-    source_pos = getGroupPosition(group_id);
+    sourcePos = getGroupPosition(groupId);
   }
   catch (GroupNotFoundException& gnfe)
   {
-    throw GroupNotFoundException("DataSet::splitGroup: group_id not found.", gnfe.getIdentifier());
+    throw GroupNotFoundException("DataSet::splitGroup: groupId not found.", gnfe.getIdentifier());
   }
-  size_t new_group_id = 0;
+  size_t newGroupId = 0;
   for (size_t i = 0; i < groups_.size(); i++)
   {
-    if (groups_[i]->getGroupId() > new_group_id)
-      new_group_id = groups_[i]->getGroupId();
+    if (groups_[i]->getGroupId() > newGroupId)
+      newGroupId = groups_[i]->getGroupId();
   }
-  new_group_id++;
-  Group new_group(new_group_id);
-  for (size_t i = 0; i < individuals_selection.size(); i++)
+  newGroupId++;
+  Group newGroup(newGroupId);
+  for (size_t i = 0; i < individualSelection.size(); i++)
   {
-    if (individuals_selection[i] >= groups_[source_pos]->getNumberOfIndividuals())
-      throw IndexOutOfBoundsException("DataSet::splitGroup: individuals_selection excedes the number of individual in the group.", individuals_selection[i], 0, groups_[source_pos]->getNumberOfIndividuals());
+    if (individualSelection[i] >= groups_[sourcePos]->getNumberOfIndividuals())
+      throw IndexOutOfBoundsException("DataSet::splitGroup: individuals_selection excedes the number of individual in the group.", individualSelection[i], 0, groups_[sourcePos]->getNumberOfIndividuals());
   }
-  for (size_t i = 0; i < individuals_selection.size(); i++)
+  for (size_t i = 0; i < individualSelection.size(); i++)
   {
-    new_group.addIndividual(*groups_[source_pos]->removeIndividualAtPosition(individuals_selection[i]));
-    groups_[source_pos]->deleteIndividualAtPosition(individuals_selection[i]);
+    newGroup.addIndividual(*groups_[sourcePos]->removeIndividualAtPosition(individualSelection[i]));
+    groups_[sourcePos]->deleteIndividualAtPosition(individualSelection[i]);
   }
-  addGroup(new_group);
+  addGroup(newGroup);
 }
 
 /******************************************************************************/
@@ -436,22 +413,22 @@ void DataSet::addEmptyIndividualToGroup(size_t group, const std::string& individ
 
 /******************************************************************************/
 
-size_t DataSet::getNumberOfIndividualsInGroup(size_t group_position) const
+size_t DataSet::getNumberOfIndividualsInGroup(size_t groupPosition) const
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::getNumberOfIndividualsInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
-  return groups_[group_position]->getNumberOfIndividuals();
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::getNumberOfIndividualsInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
+  return groups_[groupPosition]->getNumberOfIndividuals();
 }
 
 /******************************************************************************/
 
-size_t DataSet::getIndividualPositionInGroup(size_t group_position, const std::string& individual_id) const
+size_t DataSet::getIndividualPositionInGroup(size_t groupPosition, const std::string& individual_id) const
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::getIndividualPositionFromGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::getIndividualPositionFromGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    return groups_[group_position]->getIndividualPosition(individual_id);
+    return groups_[groupPosition]->getIndividualPosition(individual_id);
   }
   catch (IndividualNotFoundException& infe)
   {
@@ -461,61 +438,61 @@ size_t DataSet::getIndividualPositionInGroup(size_t group_position, const std::s
 
 /******************************************************************************/
 
-const Individual* DataSet::getIndividualAtPositionFromGroup(size_t group_position, size_t individual_position) const
+const Individual& DataSet::getIndividualAtPositionFromGroup(size_t groupPosition, size_t individualPosition) const
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::getIndividualAtPositionFromGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::getIndividualAtPositionFromGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    return &groups_[group_position]->getIndividualAtPosition(individual_position);
+    return groups_[groupPosition]->getIndividualAtPosition(individualPosition);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    throw IndexOutOfBoundsException("DataSet::getIndividualAtPositionFromGroup: individual_position out of bouds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    throw IndexOutOfBoundsException("DataSet::getIndividualAtPositionFromGroup: individualPosition out of bouds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
   }
 }
 
 /******************************************************************************/
 
-const Individual* DataSet::getIndividualByIdFromGroup(size_t group_position, const std::string& individual_id) const
+const Individual& DataSet::getIndividualByIdFromGroup(size_t groupPosition, const std::string& individualId) const
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::getIndividualByIdFromGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::getIndividualByIdFromGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    return &groups_[group_position]->getIndividualById(individual_id);
+    return groups_[groupPosition]->getIndividualById(individualId);
   }
   catch (IndividualNotFoundException& infe)
   {
-    throw IndividualNotFoundException("DataSet::getIndividualByIdFromGroup: individual_id not found.", infe.getIdentifier());
+    throw IndividualNotFoundException("DataSet::getIndividualByIdFromGroup: individualId not found.", infe.getIdentifier());
   }
 }
 
 /******************************************************************************/
 
-void DataSet::deleteIndividualAtPositionFromGroup(size_t group_position, size_t individual_position)
+void DataSet::deleteIndividualAtPositionFromGroup(size_t groupPosition, size_t individualPosition)
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::deleteIndividualAtPositionFromGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::deleteIndividualAtPositionFromGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    groups_[group_position]->deleteIndividualAtPosition(individual_position);
+    groups_[groupPosition]->deleteIndividualAtPosition(individualPosition);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    throw IndexOutOfBoundsException("DataSet::deleteIndividualAtPositionFromGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    throw IndexOutOfBoundsException("DataSet::deleteIndividualAtPositionFromGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
   }
 }
 
 /******************************************************************************/
 
-void DataSet::deleteIndividualByIdFromGroup(size_t group_position, const std::string& individual_id)
+void DataSet::deleteIndividualByIdFromGroup(size_t groupPosition, const std::string& individual_id)
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::deleteIndividualByIdFromGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::deleteIndividualByIdFromGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    groups_[group_position]->deleteIndividualById(individual_id);
+    groups_[groupPosition]->deleteIndividualById(individual_id);
   }
   catch (IndividualNotFoundException& infe)
   {
@@ -525,65 +502,65 @@ void DataSet::deleteIndividualByIdFromGroup(size_t group_position, const std::st
 
 /******************************************************************************/
 
-void DataSet::setIndividualSexInGroup(size_t group_position, size_t individual_position, const unsigned short sex)
+void DataSet::setIndividualSexInGroup(size_t groupPosition, size_t individualPosition, const unsigned short sex)
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::setIndividualSexInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::setIndividualSexInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    groups_[group_position]->setIndividualSexAtPosition(individual_position, sex);
+    groups_[groupPosition]->setIndividualSexAtPosition(individualPosition, sex);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    throw IndexOutOfBoundsException("DataSet::setIndividualSexInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    throw IndexOutOfBoundsException("DataSet::setIndividualSexInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
   }
 }
 
 /******************************************************************************/
 
-unsigned short DataSet::getIndividualSexInGroup(size_t group_position, size_t individual_position) const
+unsigned short DataSet::getIndividualSexInGroup(size_t groupPosition, size_t individualPosition) const
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::getIndividualSexInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::getIndividualSexInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    return groups_[group_position]->getIndividualSexAtPosition(individual_position);
+    return groups_[groupPosition]->getIndividualSexAtPosition(individualPosition);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    throw IndexOutOfBoundsException("DataSet::getIndividualSexInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    throw IndexOutOfBoundsException("DataSet::getIndividualSexInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
   }
 }
 
 /******************************************************************************/
 
-void DataSet::setIndividualDateInGroup(size_t group_position, size_t individual_position, const Date& date)
+void DataSet::setIndividualDateInGroup(size_t groupPosition, size_t individualPosition, const Date& date)
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::setIndividualDateInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::setIndividualDateInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    groups_[group_position]->setIndividualDateAtPosition(individual_position, date);
+    groups_[groupPosition]->setIndividualDateAtPosition(individualPosition, date);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    throw IndexOutOfBoundsException("DataSet::setIndividualDateInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    throw IndexOutOfBoundsException("DataSet::setIndividualDateInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
   }
 }
 
 /******************************************************************************/
 
-const Date* DataSet::getIndividualDateInGroup(size_t group_position, size_t individual_position) const
+const Date& DataSet::individualDateInGroup(size_t groupPosition, size_t individualPosition) const
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::getIndividualDateInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::getIndividualDateInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    return &groups_[group_position]->getIndividualDateAtPosition(individual_position);
+    return groups_[groupPosition]->getIndividualDateAtPosition(individualPosition);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    throw IndexOutOfBoundsException("DataSet::getIndividualDateInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    throw IndexOutOfBoundsException("DataSet::getIndividualDateInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
   }
   catch (NullPointerException&)
   {
@@ -593,33 +570,33 @@ const Date* DataSet::getIndividualDateInGroup(size_t group_position, size_t indi
 
 /******************************************************************************/
 
-void DataSet::setIndividualCoordInGroup(size_t group_position, size_t individual_position, const Point2D<double>& coord)
+void DataSet::setIndividualCoordInGroup(size_t groupPosition, size_t individualPosition, const Point2D<double>& coord)
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::setIndividualCoordInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::setIndividualCoordInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    groups_[group_position]->setIndividualCoordAtPosition(individual_position, coord);
+    groups_[groupPosition]->setIndividualCoordAtPosition(individualPosition, coord);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    throw IndexOutOfBoundsException("DataSet::setIndividualCoordInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    throw IndexOutOfBoundsException("DataSet::setIndividualCoordInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
   }
 }
 
 /******************************************************************************/
 
-const Point2D<double>* DataSet::getIndividualCoordInGroup(size_t group_position, size_t individual_position) const
+const Point2D<double>& DataSet::individualCoordInGroup(size_t groupPosition, size_t individualPosition) const
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::getIndividualCoordInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::getIndividualCoordInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    return &groups_[group_position]->getIndividualCoordAtPosition(individual_position);
+    return groups_[groupPosition]->getIndividualCoordAtPosition(individualPosition);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    throw IndexOutOfBoundsException("DataSet::getIndividualCoordAtPosition: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    throw IndexOutOfBoundsException("DataSet::getIndividualCoordAtPosition: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
   }
   catch (NullPointerException&)
   {
@@ -629,17 +606,17 @@ const Point2D<double>* DataSet::getIndividualCoordInGroup(size_t group_position,
 
 /******************************************************************************/
 
-void DataSet::setIndividualLocalityInGroupByName(size_t group_position, size_t individual_position, const std::string& locality_name)
+void DataSet::setIndividualLocalityInGroupByName(size_t groupPosition, size_t individualPosition, const std::string& locality_name)
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::setIndividualLocalityInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::setIndividualLocalityInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    groups_[group_position]->setIndividualLocalityAtPosition(individual_position, &getLocalityByName(locality_name));
+    groups_[groupPosition]->setIndividualLocalityAtPosition(individualPosition, getLocalityByName(locality_name));
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    throw IndexOutOfBoundsException("DataSet::setIndividualLocalityInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    throw IndexOutOfBoundsException("DataSet::setIndividualLocalityInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
   }
   catch (LocalityNotFoundException& lnfe)
   {
@@ -649,17 +626,17 @@ void DataSet::setIndividualLocalityInGroupByName(size_t group_position, size_t i
 
 /******************************************************************************/
 
-const Locality<double>* DataSet::getIndividualLocalityInGroup(size_t group_position, size_t individual_position) const
+shared_ptr<const Locality<double>> DataSet::getIndividualLocalityInGroup(size_t groupPosition, size_t individualPosition) const
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::getIndividualLocalityInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::getIndividualLocalityInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    return &groups_[group_position]->getIndividualLocalityAtPosition(individual_position);
+    return groups_[groupPosition]->getIndividualLocalityAtPosition(individualPosition);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    throw IndexOutOfBoundsException("DataSet::getIndividualLocalityInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    throw IndexOutOfBoundsException("DataSet::getIndividualLocalityInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
   }
   catch (NullPointerException&)
   {
@@ -669,22 +646,22 @@ const Locality<double>* DataSet::getIndividualLocalityInGroup(size_t group_posit
 
 /******************************************************************************/
 
-void DataSet::addIndividualSequenceInGroup(size_t group_position, size_t individual_position, size_t sequence_position, const Sequence& sequence)
+void DataSet::addIndividualSequenceInGroup(size_t groupPosition, size_t individualPosition, size_t sequence_position, unique_ptr<Sequence>& sequence)
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::addIndividualSequenceInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::addIndividualSequenceInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    groups_[group_position]->addIndividualSequenceAtPosition(individual_position, sequence_position, sequence);
-    setAlphabet(sequence.getAlphabet());
+    groups_[groupPosition]->addIndividualSequenceAtPosition(individualPosition, sequence_position, sequence);
+    setAlphabet(sequence->getAlphabet());
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    throw IndexOutOfBoundsException("DataSet::addIndividualSequenceInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    throw IndexOutOfBoundsException("DataSet::addIndividualSequenceInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
   }
   catch (AlphabetMismatchException& ame)
   {
-    throw AlphabetMismatchException("DataSet::addIndividualSequenceInGroup: sequence's alphabet doesn't match.", ame.getAlphabets()[0], ame.getAlphabets()[1]);
+    throw AlphabetMismatchException("DataSet::addIndividualSequenceInGroup: sequence's alphabet doesn't match.", ame.getFirstAlphabet(), ame.getSecondAlphabet());
   }
   catch (BadIdentifierException& bie)
   {
@@ -698,17 +675,17 @@ void DataSet::addIndividualSequenceInGroup(size_t group_position, size_t individ
 
 /******************************************************************************/
 
-const Sequence& DataSet::getIndividualSequenceByNameInGroup(size_t group_position, size_t individual_position, const std::string& sequence_name) const
+const Sequence& DataSet::getIndividualSequenceByNameInGroup(size_t groupPosition, size_t individualPosition, const std::string& sequence_name) const
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::getIndividualSequenceByNameInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::getIndividualSequenceByNameInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    return groups_[group_position]->getIndividualSequenceByName(individual_position, sequence_name);
+    return groups_[groupPosition]->getIndividualSequenceByName(individualPosition, sequence_name);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    throw IndexOutOfBoundsException("DataSet::getIndividualSequenceByNameInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    throw IndexOutOfBoundsException("DataSet::getIndividualSequenceByNameInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
   }
   catch (NullPointerException&)
   {
@@ -722,18 +699,18 @@ const Sequence& DataSet::getIndividualSequenceByNameInGroup(size_t group_positio
 
 /******************************************************************************/
 
-const Sequence& DataSet::getIndividualSequenceAtPositionInGroup(size_t group_position, size_t individual_position, size_t sequence_position) const
+const Sequence& DataSet::getIndividualSequenceAtPositionInGroup(size_t groupPosition, size_t individualPosition, size_t sequence_position) const
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::getIndividualSequenceAtPositionInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::getIndividualSequenceAtPositionInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    return groups_[group_position]->getIndividualSequenceAtPosition(individual_position, sequence_position);
+    return groups_[groupPosition]->getIndividualSequenceAtPosition(individualPosition, sequence_position);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    if (string(ioobe.what()).find("individual_position") < string(ioobe.what()).size())
-      throw IndexOutOfBoundsException("DataSet::getIndividualSequenceAtPositionInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    if (string(ioobe.what()).find("individualPosition") < string(ioobe.what()).size())
+      throw IndexOutOfBoundsException("DataSet::getIndividualSequenceAtPositionInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
     // if (string(ioobe.what()).find("sequence_position") < string(ioobe.what()).size())
     else
       throw IndexOutOfBoundsException("DataSet::getIndividualSequenceAtPositionInGroup: sequence_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
@@ -746,17 +723,17 @@ const Sequence& DataSet::getIndividualSequenceAtPositionInGroup(size_t group_pos
 
 /******************************************************************************/
 
-void DataSet::deleteIndividualSequenceByNameInGroup(size_t group_position, size_t individual_position, const std::string& sequence_name)
+void DataSet::deleteIndividualSequenceByNameInGroup(size_t groupPosition, size_t individualPosition, const std::string& sequence_name)
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::deleteIndividualSequenceByNameInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::deleteIndividualSequenceByNameInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    groups_[group_position]->deleteIndividualSequenceByName(individual_position, sequence_name);
+    groups_[groupPosition]->deleteIndividualSequenceByName(individualPosition, sequence_name);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    throw IndexOutOfBoundsException("DataSet::deleteIndividualSequenceByNameInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    throw IndexOutOfBoundsException("DataSet::deleteIndividualSequenceByNameInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
   }
   catch (NullPointerException&)
   {
@@ -770,18 +747,18 @@ void DataSet::deleteIndividualSequenceByNameInGroup(size_t group_position, size_
 
 /******************************************************************************/
 
-void DataSet::deleteIndividualSequenceAtPositionInGroup(size_t group_position, size_t individual_position, size_t sequence_position)
+void DataSet::deleteIndividualSequenceAtPositionInGroup(size_t groupPosition, size_t individualPosition, size_t sequence_position)
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::deleteIndividualSequenceAtPositionInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::deleteIndividualSequenceAtPositionInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    groups_[group_position]->deleteIndividualSequenceAtPosition(individual_position, sequence_position);
+    groups_[groupPosition]->deleteIndividualSequenceAtPosition(individualPosition, sequence_position);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    if (string(ioobe.what()).find("individual_position") < string(ioobe.what()).size())
-      throw IndexOutOfBoundsException("DataSet::deleteIndividualSequenceAtPositionInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    if (string(ioobe.what()).find("individualPosition") < string(ioobe.what()).size())
+      throw IndexOutOfBoundsException("DataSet::deleteIndividualSequenceAtPositionInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
     // if (string(ioobe.what()).find("sequence_position") < string(ioobe.what()).size())
     else
       throw IndexOutOfBoundsException("DataSet::deleteIndividualSequenceAtPositionInGroup: sequence_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
@@ -794,17 +771,17 @@ void DataSet::deleteIndividualSequenceAtPositionInGroup(size_t group_position, s
 
 /******************************************************************************/
 
-std::vector<std::string> DataSet::getIndividualSequencesNamesInGroup(size_t group_position, size_t individual_position) const
+std::vector<std::string> DataSet::getIndividualSequencesNamesInGroup(size_t groupPosition, size_t individualPosition) const
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::getIndividualSequencesNamesInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::getIndividualSequencesNamesInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    return groups_[group_position]->getIndividualSequencesNames(individual_position);
+    return groups_[groupPosition]->getIndividualSequencesNames(individualPosition);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    throw IndexOutOfBoundsException("DataSet::getIndividualSequencesNamesInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    throw IndexOutOfBoundsException("DataSet::getIndividualSequencesNamesInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
   }
   catch (NullPointerException&)
   {
@@ -814,17 +791,17 @@ std::vector<std::string> DataSet::getIndividualSequencesNamesInGroup(size_t grou
 
 /******************************************************************************/
 
-size_t DataSet::getIndividualSequencePositionInGroup(size_t group_position, size_t individual_position, const std::string& sequence_name) const
+size_t DataSet::getIndividualSequencePositionInGroup(size_t groupPosition, size_t individualPosition, const std::string& sequence_name) const
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::getIndividualSequencePositionInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::getIndividualSequencePositionInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    return groups_[group_position]->getIndividualSequencePosition(individual_position, sequence_name);
+    return groups_[groupPosition]->getIndividualSequencePosition(individualPosition, sequence_name);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    throw IndexOutOfBoundsException("DataSet::getIndividualSequencePositionInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    throw IndexOutOfBoundsException("DataSet::getIndividualSequencePositionInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
   }
   catch (NullPointerException&)
   {
@@ -838,17 +815,17 @@ size_t DataSet::getIndividualSequencePositionInGroup(size_t group_position, size
 
 /******************************************************************************/
 
-size_t DataSet::getIndividualNumberOfSequencesInGroup(size_t group_position, size_t individual_position) const
+size_t DataSet::getIndividualNumberOfSequencesInGroup(size_t groupPosition, size_t individualPosition) const
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::getIndividualNumberOfSequencesInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::getIndividualNumberOfSequencesInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    return groups_[group_position]->getIndividualNumberOfSequences(individual_position);
+    return groups_[groupPosition]->getIndividualNumberOfSequences(individualPosition);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    throw IndexOutOfBoundsException("DataSet::getIndividualNumberOfSequencesInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    throw IndexOutOfBoundsException("DataSet::getIndividualNumberOfSequencesInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
   }
   catch (NullPointerException&)
   {
@@ -858,33 +835,33 @@ size_t DataSet::getIndividualNumberOfSequencesInGroup(size_t group_position, siz
 
 /******************************************************************************/
 
-void DataSet::setIndividualGenotypeInGroup(size_t group_position, size_t individual_position, const MultilocusGenotype& genotype)
+void DataSet::setIndividualGenotypeInGroup(size_t groupPosition, size_t individualPosition, const MultilocusGenotype& genotype)
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::setIndividualGenotypeInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::setIndividualGenotypeInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    groups_[group_position]->setIndividualGenotype(individual_position, genotype);
+    groups_[groupPosition]->setIndividualGenotype(individualPosition, genotype);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    throw IndexOutOfBoundsException("DataSet::setIndividualGenotypeInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    throw IndexOutOfBoundsException("DataSet::setIndividualGenotypeInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
   }
 }
 
 /******************************************************************************/
 
-void DataSet::initIndividualGenotypeInGroup(size_t group_position, size_t individual_position)
+void DataSet::initIndividualGenotypeInGroup(size_t groupPosition, size_t individualPosition)
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::initIndividualGenotypeInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::initIndividualGenotypeInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    groups_[group_position]->initIndividualGenotype(individual_position, getNumberOfLoci());
+    groups_[groupPosition]->initIndividualGenotype(individualPosition, getNumberOfLoci());
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    throw IndexOutOfBoundsException("DataSet::initIndividualGenotypeInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    throw IndexOutOfBoundsException("DataSet::initIndividualGenotypeInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
   }
   catch (BadIntegerException& bie)
   {
@@ -902,34 +879,38 @@ void DataSet::initIndividualGenotypeInGroup(size_t group_position, size_t indivi
 
 /******************************************************************************/
 
-void DataSet::deleteIndividualGenotypeInGroup(size_t group_position, size_t individual_position)
+void DataSet::deleteIndividualGenotypeInGroup(size_t groupPosition, size_t individualPosition)
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::deleteIndividualGenotypeInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::deleteIndividualGenotypeInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    groups_[group_position]->deleteIndividualGenotype(individual_position);
+    groups_[groupPosition]->deleteIndividualGenotype(individualPosition);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    throw IndexOutOfBoundsException("DataSet::deleteIndividualGenotypeInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    throw IndexOutOfBoundsException("DataSet::deleteIndividualGenotypeInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
   }
 }
 
 /******************************************************************************/
 
-void DataSet::setIndividualMonolocusGenotypeInGroup(size_t group_position, size_t individual_position, size_t locus_position, const MonolocusGenotype& monogen)
+void DataSet::setIndividualMonolocusGenotypeInGroup(
+    size_t groupPosition,
+    size_t individualPosition,
+    size_t locusPosition,
+    const MonolocusGenotypeInterface& monogen)
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::setIndividualMonolocusGenotypeInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::setIndividualMonolocusGenotypeInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    groups_[group_position]->setIndividualMonolocusGenotype(individual_position, locus_position, monogen);
+    groups_[groupPosition]->setIndividualMonolocusGenotype(individualPosition, locusPosition, monogen);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    if (string(ioobe.what()).find("individual_position") < string(ioobe.what()).size())
-      throw IndexOutOfBoundsException("DataSet::setIndividualMonolocusGenotypeInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    if (string(ioobe.what()).find("individualPosition") < string(ioobe.what()).size())
+      throw IndexOutOfBoundsException("DataSet::setIndividualMonolocusGenotypeInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
     // if (string(ioobe.what()).find("locus_position") < string(ioobe.what()).size())
     else
       throw IndexOutOfBoundsException("DataSet::setIndividualMonolocusGenotypeInGroup: locus_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
@@ -942,18 +923,18 @@ void DataSet::setIndividualMonolocusGenotypeInGroup(size_t group_position, size_
 
 /******************************************************************************/
 
-void DataSet::setIndividualMonolocusGenotypeByAlleleKeyInGroup(size_t group_position, size_t individual_position, size_t locus_position, const std::vector<size_t> allele_keys)
+void DataSet::setIndividualMonolocusGenotypeByAlleleKeyInGroup(size_t groupPosition, size_t individualPosition, size_t locus_position, const std::vector<size_t> allele_keys)
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::setIndividualMonolocusGenotypeByAlleleKeyInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::setIndividualMonolocusGenotypeByAlleleKeyInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    groups_[group_position]->setIndividualMonolocusGenotypeByAlleleKey(individual_position, locus_position, allele_keys);
+    groups_[groupPosition]->setIndividualMonolocusGenotypeByAlleleKey(individualPosition, locus_position, allele_keys);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    if (string(ioobe.what()).find("individual_position") < string(ioobe.what()).size())
-      throw IndexOutOfBoundsException("DataSet::setIndividualMonolocusGenotypeByAlleleKeyInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    if (string(ioobe.what()).find("individualPosition") < string(ioobe.what()).size())
+      throw IndexOutOfBoundsException("DataSet::setIndividualMonolocusGenotypeByAlleleKeyInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
     // if (string(ioobe.what()).find("locus_position") < string(ioobe.what()).size())
     else
       throw IndexOutOfBoundsException("DataSet::setIndividualMonolocusGenotypeByAlleleKeyInGroup: locus_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
@@ -970,19 +951,19 @@ void DataSet::setIndividualMonolocusGenotypeByAlleleKeyInGroup(size_t group_posi
 
 /******************************************************************************/
 
-void DataSet::setIndividualMonolocusGenotypeByAlleleIdInGroup(size_t group_position, size_t individual_position, size_t locus_position, const std::vector<std::string> allele_id)
+void DataSet::setIndividualMonolocusGenotypeByAlleleIdInGroup(size_t groupPosition, size_t individualPosition, size_t locus_position, const std::vector<std::string> allele_id)
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::setIndividualMonolocusGenotypeByAlleleIdInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::setIndividualMonolocusGenotypeByAlleleIdInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   const LocusInfo& locus_info = getLocusInfoAtPosition(locus_position);
   try
   {
-    groups_[group_position]->setIndividualMonolocusGenotypeByAlleleId(individual_position, locus_position, allele_id, locus_info);
+    groups_[groupPosition]->setIndividualMonolocusGenotypeByAlleleId(individualPosition, locus_position, allele_id, locus_info);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    if (string(ioobe.what()).find("individual_position") < string(ioobe.what()).size())
-      throw IndexOutOfBoundsException("DataSet::setIndividualMonolocusGenotypeByAlleleIdInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    if (string(ioobe.what()).find("individualPosition") < string(ioobe.what()).size())
+      throw IndexOutOfBoundsException("DataSet::setIndividualMonolocusGenotypeByAlleleIdInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
     // if (string(ioobe.what()).find("locus_position") < string(ioobe.what()).size())
     else
       throw IndexOutOfBoundsException("DataSet::setIndividualMonolocusGenotypeByAlleleIdInGroup: locus_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
@@ -999,18 +980,21 @@ void DataSet::setIndividualMonolocusGenotypeByAlleleIdInGroup(size_t group_posit
 
 /******************************************************************************/
 
-const MonolocusGenotype* DataSet::getIndividualMonolocusGenotypeInGroup(size_t group_position, size_t individual_position, size_t locus_position) const
+const MonolocusGenotypeInterface& DataSet::getIndividualMonolocusGenotypeInGroup(
+    size_t groupPosition,
+    size_t individualPosition,
+    size_t locusPosition) const
 {
-  if (group_position >= getNumberOfGroups())
-    throw IndexOutOfBoundsException("DataSet::getIndividualMonolocusGenotypeInGroup: group_position out of bounds.", group_position, 0, getNumberOfGroups());
+  if (groupPosition >= getNumberOfGroups())
+    throw IndexOutOfBoundsException("DataSet::getIndividualMonolocusGenotypeInGroup: groupPosition out of bounds.", groupPosition, 0, getNumberOfGroups());
   try
   {
-    return &groups_[group_position]->getIndividualMonolocusGenotype(individual_position, locus_position);
+    return groups_[groupPosition]->getIndividualMonolocusGenotype(individualPosition, locusPosition);
   }
   catch (IndexOutOfBoundsException& ioobe)
   {
-    if (string(ioobe.what()).find("individual_position") < string(ioobe.what()).size())
-      throw IndexOutOfBoundsException("DataSet::getIndividualMonolocusGenotypeInGroup: individual_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
+    if (string(ioobe.what()).find("individualPosition") < string(ioobe.what()).size())
+      throw IndexOutOfBoundsException("DataSet::getIndividualMonolocusGenotypeInGroup: individualPosition out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
     // if (string(ioobe.what()).find("locus_position") < string(ioobe.what()).size())
     else
       throw IndexOutOfBoundsException("DataSet::getIndividualMonolocusGenotypeInGroup: locus_position out of bounds.", ioobe.getBadIndex(), ioobe.getBounds()[0], ioobe.getBounds()[1]);
@@ -1023,86 +1007,16 @@ const MonolocusGenotype* DataSet::getIndividualMonolocusGenotypeInGroup(size_t g
 
 /******************************************************************************/
 
-// Dealing with AnalyzedSequences --------------------------
-
-void DataSet::setAlphabet(const Alphabet* alpha)
+void DataSet::setAlphabet(const std::string& alphaType)
 {
-  if (analyzedSequences_ == 0)
-    analyzedSequences_ = new AnalyzedSequences();
-  analyzedSequences_->setAlphabet(alpha);
-}
-
-/******************************************************************************/
-
-void DataSet::setAlphabet(const std::string& alpha_type)
-{
-  if (analyzedSequences_ == 0)
-    analyzedSequences_ = new AnalyzedSequences();
-  analyzedSequences_->setAlphabet(alpha_type);
-}
-
-/******************************************************************************/
-
-const Alphabet* DataSet::getAlphabet() const
-{
-  if (analyzedSequences_ != 0)
-    return analyzedSequences_->getAlphabet();
-  throw NullPointerException("DataSet::getAlphabet: no sequence data.");
-}
-
-/******************************************************************************/
-
-std::string DataSet::getAlphabetType() const
-{
-  if (analyzedSequences_ != 0)
-    return analyzedSequences_->getAlphabetType();
-  throw NullPointerException("DataSet::getAlphabetType: no sequence data.");
-}
-
-/******************************************************************************/
-
-// Dealing with AnalyzedLoci -------------------------------
-
-void DataSet::setAnalyzedLoci(const AnalyzedLoci& analyzedLoci)
-{
-  if (analyzedLoci_ != 0)
-  {
-    try
-    {
-      deleteAnalyzedLoci();
-    }
-    catch (Exception& e)
-    {
-      throw Exception ("DataSet::setAnalyzedLoci: at least one individual has a genotype of the actual AnalyzedLoci.");
-    }
-  }
-  analyzedLoci_ = new AnalyzedLoci(analyzedLoci);
-}
-
-/******************************************************************************/
-
-void DataSet::initAnalyzedLoci(size_t number_of_loci)
-{
-  if (analyzedLoci_ != 0)
-    throw Exception("DataSet::initAnalyzedLoci: analyzedLoci_ already initialyzed.");
-  analyzedLoci_ = new AnalyzedLoci(number_of_loci);
-}
-
-/******************************************************************************/
-
-const AnalyzedLoci* DataSet::getAnalyzedLoci() const
-{
-  if (analyzedLoci_ != 0)
-    return analyzedLoci_;
-  throw NullPointerException("DataSet::getAnalyzedLoci: no loci initialized.");
-}
-
-/******************************************************************************/
-
-void DataSet::deleteAnalyzedLoci()
-{
-  if (analyzedLoci_ != 0)
-    delete analyzedLoci_;
+  if (alphaType != string("DNA") && alphaType != string("RNA") && alphaType != string("PROTEIN"))
+    throw Exception(string("DataSet::setAlphabet: bad alphabet type. (") + alphaType + string(")."));
+  if (alphaType == string("DNA"))
+    sequenceAlphabet_ = AlphabetTools::DNA_ALPHABET;
+  if (alphaType == string("RNA"))
+    sequenceAlphabet_ = AlphabetTools::RNA_ALPHABET;
+  if (alphaType == string("PROTEIN"))
+    sequenceAlphabet_ = AlphabetTools::PROTEIN_ALPHABET;
 }
 
 /******************************************************************************/
@@ -1240,23 +1154,20 @@ size_t DataSet::getPloidyByLocusPosition(size_t locus_position) const
 
 /******************************************************************************/
 
-// Container extraction -----------------------------------
-
-PolymorphismMultiGContainer* DataSet::getPolymorphismMultiGContainer() const
+unique_ptr<PolymorphismMultiGContainer> DataSet::getPolymorphismMultiGContainer() const
 {
-  PolymorphismMultiGContainer* pmgc = new PolymorphismMultiGContainer();
+  auto pmgc = make_unique<PolymorphismMultiGContainer>();
   for (size_t i = 0; i < getNumberOfGroups(); i++)
   {
-    // nommer les groupes khalid
     string name = groups_[i]->getGroupName();
     pmgc->addGroupName(i, name);
-    for (size_t j = 0; j < getNumberOfIndividualsInGroup(i); j++)
+    for (size_t j = 0; j < getNumberOfIndividualsInGroup(i); ++j)
     {
-      const Individual* tmp_ind = getIndividualAtPositionFromGroup(i, j);
-      if (tmp_ind->hasGenotype())
+      const auto& tmpInd = getIndividualAtPositionFromGroup(i, j);
+      if (tmpInd.hasGenotype())
       {
-        const MultilocusGenotype& tmp_mg = tmp_ind->getGenotype();
-        pmgc->addMultilocusGenotype(tmp_mg, i);
+        auto tmpMg = unique_ptr<MultilocusGenotype>(tmpInd.getGenotype().clone());
+        pmgc->addMultilocusGenotype(tmpMg, i);
       }
     }
   }
@@ -1265,15 +1176,15 @@ PolymorphismMultiGContainer* DataSet::getPolymorphismMultiGContainer() const
 
 /******************************************************************************/
 
-PolymorphismMultiGContainer* DataSet::getPolymorphismMultiGContainer(const std::map<size_t, std::vector<size_t> >& selection) const
+unique_ptr<PolymorphismMultiGContainer> DataSet::getPolymorphismMultiGContainer(const std::map<size_t, std::vector<size_t> >& selection) const
 {
-  PolymorphismMultiGContainer* pmgc = new PolymorphismMultiGContainer();
-  for (map<size_t, vector<size_t> >::const_iterator it = selection.begin(); it != selection.end(); it++)
+  auto pmgc = make_unique<PolymorphismMultiGContainer>();
+  for (const auto& it : selection)
   {
     size_t i;
     try
     {
-      i = getGroupPosition(it->first);
+      i = getGroupPosition(it.first);
     }
     catch (GroupNotFoundException& gnfe)
     {
@@ -1281,21 +1192,20 @@ PolymorphismMultiGContainer* DataSet::getPolymorphismMultiGContainer(const std::
     }
     string name = groups_[i]->getGroupName();
     pmgc->addGroupName(i, name);
-    for (size_t j = 0; j < it->second.size(); j++)
+    for (size_t j = 0; j < it.second.size(); j++)
     {
-      const Individual* tmp_ind = 0;
       try
       {
-        tmp_ind = getIndividualAtPositionFromGroup(i, j);
+        const auto& tmpInd = getIndividualAtPositionFromGroup(i, j);
+        if (tmpInd.hasGenotype())
+        {
+          auto tmpMg = unique_ptr<MultilocusGenotype>(tmpInd.getGenotype().clone());
+          pmgc->addMultilocusGenotype(tmpMg, i);
+        }
       }
       catch (IndexOutOfBoundsException& ioobe)
       {
         throw ioobe;
-      }
-      if (tmp_ind->hasGenotype())
-      {
-        const MultilocusGenotype& tmp_mg = tmp_ind->getGenotype();
-        pmgc->addMultilocusGenotype(tmp_mg, i);
       }
     }
   }
@@ -1304,38 +1214,37 @@ PolymorphismMultiGContainer* DataSet::getPolymorphismMultiGContainer(const std::
 
 /******************************************************************************/
 
-PolymorphismSequenceContainer* DataSet::getPolymorphismSequenceContainer(const std::map<size_t, std::vector<size_t> >& selection, size_t sequence_position) const
+unique_ptr<PolymorphismSequenceContainer> DataSet::getPolymorphismSequenceContainer(
+    const std::map<size_t, std::vector<size_t> >& selection,
+    size_t sequencePosition) const
 {
-  PolymorphismSequenceContainer* psc = new PolymorphismSequenceContainer(getAlphabet());
-  for (map<size_t, vector<size_t> >::const_iterator it = selection.begin(); it != selection.end(); it++)
+  auto psc = make_unique<PolymorphismSequenceContainer>(getAlphabet());
+  for (auto& it : selection)
   {
     size_t i;
     try
     {
-      i = getGroupPosition(it->first);
+      i = getGroupPosition(it.first);
     }
     catch (GroupNotFoundException& gnfe)
     {
-      delete psc;
       throw gnfe;
     }
-    for (size_t j = 0; j < it->second.size(); j++)
+    for (size_t j = 0; j < it.second.size(); ++j)
     {
-      const Individual* tmp_ind = 0;
       try
       {
-        tmp_ind = getIndividualAtPositionFromGroup(i, j);
+        const auto& tmpInd = getIndividualAtPositionFromGroup(i, j);
+        if (tmpInd.hasSequenceAtPosition(sequencePosition))
+        {
+	  auto tmpSeq = unique_ptr<Sequence>(tmpInd.sequenceAtPosition(sequencePosition).clone());
+          psc->addSequenceWithFrequency(tmpSeq->getName(), tmpSeq, 1);
+          psc->setGroupId(tmpSeq->getName(), it.first);
+        }
       }
       catch (IndexOutOfBoundsException& ioobe)
       {
-        delete psc;
         throw ioobe;
-      }
-      if (tmp_ind->hasSequenceAtPosition(sequence_position))
-      {
-        const Sequence& tmp_seq = tmp_ind->getSequenceAtPosition(sequence_position);
-        psc->addSequenceWithFrequency(tmp_seq, 1, false);
-        psc->setGroupId(tmp_seq.getName(), it->first);
       }
     }
   }
@@ -1344,18 +1253,3 @@ PolymorphismSequenceContainer* DataSet::getPolymorphismSequenceContainer(const s
 
 /******************************************************************************/
 
-// General tests ------------------------------------------
-
-bool DataSet::hasSequenceData() const
-{
-  return analyzedSequences_ != 0;
-}
-
-/******************************************************************************/
-
-bool DataSet::hasAlleleicData() const
-{
-  return analyzedLoci_ != 0;
-}
-
-/******************************************************************************/
