@@ -60,17 +60,16 @@ namespace bpp
  * @author Sylvain Gaillard
  */
 class MultiAlleleMonolocusGenotype :
-  public MonolocusGenotype
+  public virtual MonolocusGenotypeInterface
 {
 private:
-  std::vector<size_t> allele_index_;
+  std::vector<size_t> alleleIndex_;
 
 public:
-  // Constructors and destructor
   /**
    * @brief Build a monolocus genotype containing many alleles.
    */
-  MultiAlleleMonolocusGenotype(std::vector<size_t> allele_index);
+  MultiAlleleMonolocusGenotype(std::vector<size_t> alleleIndex);
 
   /**
    * @brief Copy constructor.
@@ -80,7 +79,7 @@ public:
   /**
    * @brief Destroy the MultiAlleleMonolocusGenotype.
    */
-  ~MultiAlleleMonolocusGenotype();
+  virtual ~MultiAlleleMonolocusGenotype();
 
 public:
   // Other methodes
@@ -104,7 +103,11 @@ public:
    *
    * @{
    */
-  std::vector<size_t> getAlleleIndex() const;
+  std::vector<size_t> getAlleleIndex() const override
+  {
+    return alleleIndex_;
+  }
+
   /** @} */
 
   /**
@@ -112,7 +115,10 @@ public:
    *
    * @{
    */
-  MultiAlleleMonolocusGenotype* clone() const;
+  MultiAlleleMonolocusGenotype* clone() const override
+  {
+    return new MultiAlleleMonolocusGenotype(*this);
+  }
   /** @} */
 };
 } // end of namespace bpp;
