@@ -50,10 +50,10 @@
 
 #include <Bpp/Exceptions.h>
 
-// From SeqLib
+// From bpp-seq
 #include <Bpp/Seq/DistanceMatrix.h>
 
-// From popgenlib
+// From bpp-popgen
 #include "PolymorphismMultiGContainer.h"
 #include "MultilocusGenotype.h"
 #include "GeneralExceptions.h"
@@ -86,77 +86,104 @@ public:
 
   struct PermResults
   {
-    double Statistic;
-    double Percent_sup;
-    double Percent_inf;
+    double statistic;
+    double percentSup;
+    double percentInf;
   };
 
   /**
    * @brief Get the alleles' id at one locus for a set of groups.
    *
-   * @throw IndexOutOfBoundsException if locus_position excedes the number of loci of one MultilocusGenotype.
+   * @throw IndexOutOfBoundsException if locusPosition excedes the number of loci of one MultilocusGenotype.
    */
-  static std::vector<size_t> getAllelesIdsForGroups(const PolymorphismMultiGContainer& pmgc, size_t locus_position, const std::set<size_t>& groups);
+  static std::vector<size_t> getAllelesIdsForGroups(
+      const PolymorphismMultiGContainer& pmgc,
+      size_t locusPosition,
+      const std::set<size_t>& groups);
 
   /**
    * @brief Count the number of allele (gametes) at a locus for a set of groups.
    *
-   * @throw IndexOutOfBoundsException if locus_position excedes the number of loci of one MultilocusGenotype.
+   * @throw IndexOutOfBoundsException if locusPosition excedes the number of loci of one MultilocusGenotype.
    */
-  static size_t countGametesForGroups(const PolymorphismMultiGContainer& pmgc, size_t locus_position, const std::set<size_t>& groups);
+  static size_t countGametesForGroups(
+      const PolymorphismMultiGContainer& pmgc,
+      size_t locusPosition,
+      const std::set<size_t>& groups);
 
   /**
    * @brief Get a map of allele count for a set of groups.
    *
-   * @throw IndexOutOfBoundsException if locus_position excedes the number of loci of one MultilocusGenotype.
+   * @throw IndexOutOfBoundsException if locusPosition excedes the number of loci of one MultilocusGenotype.
    */
-  static std::map<size_t, size_t> getAllelesMapForGroups(const PolymorphismMultiGContainer& pmgc, size_t locus_position, const std::set<size_t>& groups);
+  static std::map<size_t, size_t> getAllelesMapForGroups(
+      const PolymorphismMultiGContainer& pmgc,
+      size_t locusPosition,
+      const std::set<size_t>& groups);
 
   /**
    * @brief Get the alleles frequencies at one locus for a set of groups.
    *
-   * @throw IndexOutOfBoundsException if locus_position excedes the number of loci of one MultilocusGenotype.
+   * @throw IndexOutOfBoundsException if locusPosition excedes the number of loci of one MultilocusGenotype.
    * @throw ZeroDivisionException if the number of considered alleles = 0.
    */
-  static std::map<size_t, double> getAllelesFrqForGroups(const PolymorphismMultiGContainer& pmgc, size_t locus_position, const std::set<size_t>& groups);
+  static std::map<size_t, double> getAllelesFrqForGroups(
+      const PolymorphismMultiGContainer& pmgc,
+      size_t locusPosition,
+      const std::set<size_t>& groups);
 
   /**
    * @brief Count the number of non-missing data at a given locus for a set of groups.
    *
-   * @throw IndexOutOfBoundsException if locus_position excedes the number of loci of one MultilocusGenotype.
+   * @throw IndexOutOfBoundsException if locusPosition excedes the number of loci of one MultilocusGenotype.
    */
-  static size_t countNonMissingForGroups(const PolymorphismMultiGContainer& pmgc, size_t locus_position, const std::set<size_t>& groups);
+  static size_t countNonMissingForGroups(
+      const PolymorphismMultiGContainer& pmgc,
+      size_t locusPosition,
+      const std::set<size_t>& groups);
 
   /**
    * @brief Counr the number of bi-allelic MonolocusGenotype at a given locus for a set of groups.
    *
-   * @throw IndexOutOfBoundsException if locus_position excedes the number of loci of one MultilocusGenotype.
+   * @throw IndexOutOfBoundsException if locusPosition excedes the number of loci of one MultilocusGenotype.
    */
-  static size_t countBiAllelicForGroups(const PolymorphismMultiGContainer& pmgc, size_t locus_position, const std::set<size_t>& groups);
+  static size_t countBiAllelicForGroups(
+      const PolymorphismMultiGContainer& pmgc,
+      size_t locusPosition,
+      const std::set<size_t>& groups);
 
   /**
    * @brief Count how many times each allele is found in an heterozygous MonolocusGenotype in a set of groups.
    *
-   * @throw IndexOutOfBoundsException if locus_position excedes the number of loci of one MultilocusGenotype.
+   * @throw IndexOutOfBoundsException if locusPosition excedes the number of loci of one MultilocusGenotype.
    */
-  static std::map<size_t, size_t> countHeterozygousForGroups(const PolymorphismMultiGContainer& pmgc, size_t locus_position, const std::set<size_t>& groups);
+  static std::map<size_t, size_t> countHeterozygousForGroups(
+      const PolymorphismMultiGContainer& pmgc,
+      size_t locusPosition,
+      const std::set<size_t>& groups);
 
   /**
    * @brief Get the heterozygous frequencies for each allele at a locus in a set of groups.
    *
-   * @throw IndexOutOfBoundsException if locus_position excedes the number of loci of one MultilocusGenotype.
+   * @throw IndexOutOfBoundsException if locusPosition excedes the number of loci of one MultilocusGenotype.
    * @throw ZeroDivisionException if the number of considered alleles = 0.
    */
-  static std::map<size_t, double> getHeterozygousFrqForGroups(const PolymorphismMultiGContainer& pmgc, size_t locus_position, const std::set<size_t>& groups);
+  static std::map<size_t, double> getHeterozygousFrqForGroups(
+      const PolymorphismMultiGContainer& pmgc, 
+      size_t locusPosition, 
+      const std::set<size_t>& groups);
 
   /**
    * @brief Compute the observed heterozygosity for one locus.
    *
    * This is the mean value of the getHeterozygousFrqForGroups map.
-   * @throw IndexOutOfBoundsException if locus_position excedes the number of loci of one MultilocusGenotype.
+   * @throw IndexOutOfBoundsException if locusPosition excedes the number of loci of one MultilocusGenotype.
    * @throw ZeroDivisionException if the number of considered alleles = 0.
    */
-  static double getHobsForGroups(const PolymorphismMultiGContainer& pmgc, size_t locus_position, const std::set<size_t>& groups);
+  static double getHobsForGroups(
+      const PolymorphismMultiGContainer& pmgc,
+      size_t locusPosition,
+      const std::set<size_t>& groups);
 
   /**
    * @brief Compute the expected heterozygosity for one locus.
@@ -166,10 +193,13 @@ public:
    * H_{exp}=1-\sum_{i=1}^{n}x_i^2
    * @f]
    * where @f$x_i@f$ is the frequency of the i<sup>th</sup> allele and @f$n@f$ the number of alleles.
-   * @throw IndexOutOfBoundsException if locus_position excedes the number of loci of one MultilocusGenotype.
+   * @throw IndexOutOfBoundsException if locusPosition excedes the number of loci of one MultilocusGenotype.
    * @throw ZeroDivisionException if the number of considered alleles = 0.
    */
-  static double getHexpForGroups(const PolymorphismMultiGContainer& pmgc, size_t locus_position, const std::set<size_t>& groups);
+  static double getHexpForGroups(
+      const PolymorphismMultiGContainer& pmgc, 
+      size_t locusPosition, 
+      const std::set<size_t>& groups);
 
   /**
    * @brief Compute the expected non biased heterozygosity for one locus.
@@ -179,10 +209,13 @@ public:
    * H_{nb}=\frac{2n}{2n-1}\left(1-\sum_{i=1}^{n}x_i^2\right)=\frac{2n}{2n-1}H_{exp}
    * @f]
    * where @f$x_i@f$ is the frequency of the i<sup>th</sup> allele and @f$n@f$ the number of alleles.
-   * @throw IndexOutOfBoundsException if locus_position excedes the number of loci of one MultilocusGenotype.
+   * @throw IndexOutOfBoundsException if locusPosition excedes the number of loci of one MultilocusGenotype.
    * @throw ZeroDivisionException if the number of considered alleles = 0.
    */
-  static double getHnbForGroups(const PolymorphismMultiGContainer& pmgc, size_t locus_position, const std::set<size_t>& groups);
+  static double getHnbForGroups(
+      const PolymorphismMultiGContainer& pmgc,
+      size_t locusPosition,
+      const std::set<size_t>& groups);
 
   /**
    * @brief Compute the Nei distance between two groups at one locus.
@@ -194,10 +227,14 @@ public:
    * @f]
    * where @f$x_i@f$ and @f$y_i@f$ are respectively the i<sup>th</sup> allele's frequency of the first and second group
    * and @f$n@f$ the total number of alleles of both groups.
-   * @throw IndexOutOfBoundsException if locus_position excedes the number of loci of one MultilocusGenotype.
+   * @throw IndexOutOfBoundsException if locusPosition excedes the number of loci of one MultilocusGenotype.
    * @throw ZeroDivisionException if the number of considered alleles = 0.
    */
-  static double getDnei72(const PolymorphismMultiGContainer& pmgc, std::vector<size_t> locus_positions, size_t grp1, size_t grp2);
+  static double getDnei72(
+      const PolymorphismMultiGContainer& pmgc, 
+      std::vector<size_t> locusPositions,
+      size_t grp1,
+      size_t grp2);
 
   /**
    * @brief Compute the Nei unbiased distance between two groups at a given number of loci.
@@ -216,47 +253,72 @@ public:
    * \qquad\textrm{and}\qquad
    * J_Y=\sum_{i=1}^{n}y_i^2
    * @f]
-   * @throw IndexOutOfBoundsException if locus_position excedes the number of loci of one MultilocusGenotype.
+   * @throw IndexOutOfBoundsException if locusPosition excedes the number of loci of one MultilocusGenotype.
    * @throw ZeroDivisionException if the number of considered alleles = 0.
    */
-  static double getDnei78(const PolymorphismMultiGContainer& pmgc, std::vector<size_t> locus_positions, size_t grp1, size_t grp2);
+  static double getDnei78(
+      const PolymorphismMultiGContainer& pmgc,
+      std::vector<size_t> locusPositions,
+      size_t grp1,
+      size_t grp2);
 
   /**
    * @brief Compute the three F statistics of Weir and Cockerham for each allele of a given locus.
    */
-  static std::map<size_t, Fstats>  getAllelesFstats(const PolymorphismMultiGContainer& pmgc, size_t locus_position, const std::set<size_t>& groups);
+  static std::map<size_t, Fstats> getAllelesFstats(
+      const PolymorphismMultiGContainer& pmgc,
+      size_t locusPosition,
+      const std::set<size_t>& groups);
 
   /**
    * @brief Compute the Weir and Cockerham Fit on a set of groups for each allele of a given locus.
    */
-  static std::map<size_t, double> getAllelesFit(const PolymorphismMultiGContainer& pmgc, size_t locus_position, const std::set<size_t>& groups);
+  static std::map<size_t, double> getAllelesFit(
+      const PolymorphismMultiGContainer& pmgc, 
+      size_t locusPosition, 
+      const std::set<size_t>& groups);
 
   /**
    * @brief Compute the Weir and Cockerham @f$\theta@f$ on a set of groups for each allele of a given locus.
    */
-  static std::map<size_t, double> getAllelesFst(const PolymorphismMultiGContainer& pmgc, size_t locus_position, const std::set<size_t>& groups);
+  static std::map<size_t, double> getAllelesFst(
+      const PolymorphismMultiGContainer& pmgc, 
+      size_t locusPosition,
+      const std::set<size_t>& groups);
 
   /**
    * @brief Compute the Weir and Cockerham Fis on a set of groups for each allele of a given locus.
    */
-  static std::map<size_t, double> getAllelesFis(const PolymorphismMultiGContainer& pmgc, size_t locus_position, const std::set<size_t>& groups);
+  static std::map<size_t, double> getAllelesFis(
+      const PolymorphismMultiGContainer& pmgc, 
+      size_t locusPosition, 
+      const std::set<size_t>& groups);
 
   /**
    * @brief Get the variance components a, b and c (Weir and Cockerham, 1983).
    */
-  static std::map<size_t, VarComp> getVarianceComponents(const PolymorphismMultiGContainer& pmgc, size_t locus_position, const std::set<size_t>& groups);
+  static std::map<size_t, VarComp> getVarianceComponents(
+      const PolymorphismMultiGContainer& pmgc, 
+      size_t locusPosition,
+      const std::set<size_t>& groups);
 
   /**
    * @brief Compute the Weir and Cockerham @f$\theta{wc}@f$ on a set of groups for a given set of loci.
    * The variance componenets for each allele are calculated and then combined over loci using Weir and Cockerham weighting.
    */
-  static double getWCMultilocusFst(const PolymorphismMultiGContainer& pmgc, std::vector<size_t> locus_positions, const std::set<size_t>& groups);
+  static double getWCMultilocusFst(
+      const PolymorphismMultiGContainer& pmgc,
+      std::vector<size_t> locusPositions, 
+      const std::set<size_t>& groups);
 
   /**
    * @brief Compute the Weir and Cockerham Fis on a set of groups for a given set of loci.
    * The variance componenets for each allele are calculated and then combined over loci using Weir and Cockerham weighting.
    */
-  static double getWCMultilocusFis(const PolymorphismMultiGContainer& pmgc, std::vector<size_t> locus_positions, const std::set<size_t>& groups);
+  static double getWCMultilocusFis(
+      const PolymorphismMultiGContainer& pmgc, 
+      std::vector<size_t> locusPositions,
+      const std::set<size_t>& groups);
 
   /**
    * @brief Compute the Weir and Cockerham @f$\theta_{wc}@f$ on a set of groups for a given set of loci and make a permutation test.
@@ -264,7 +326,11 @@ public:
    * a permutation of individuals between the different groups.
    * Return values are theta, % of values > theta and % of values < theta.
    */
-  static PermResults getWCMultilocusFstAndPerm(const PolymorphismMultiGContainer& pmgc, std::vector<size_t> locus_positions, std::set<size_t> groups, int nb_perm);
+  static PermResults getWCMultilocusFstAndPerm(
+      const PolymorphismMultiGContainer& pmgc, 
+      std::vector<size_t> locusPositions, 
+      std::set<size_t> groups, 
+      unsigned int nb_perm);
 
   /**
    * @brief Compute the Weir and Cockerham Fis on a set of groups for a given set of loci and make a permutation test.
@@ -272,21 +338,32 @@ public:
    * a permutation of alleles between individual of each group.
    * Return values are Fis, % of values > Fis and % of values < Fis.
    */
-  static PermResults getWCMultilocusFisAndPerm(const PolymorphismMultiGContainer& pmgc, std::vector<size_t> locus_positions, std::set<size_t> groups, int nb_perm);
+  static PermResults getWCMultilocusFisAndPerm(
+      const PolymorphismMultiGContainer& pmgc, 
+      std::vector<size_t> locusPositions,
+      std::set<size_t> groups,
+      unsigned int nbPerm);
 
 
   /**
    * @brief Compute the @f$\theta_{RH}@f$ on a set of groups for a given set of loci.
    * The variance componenets for each allele are calculated and then combined over loci using RH weighting with alleles frequency.
    */
-  static double getRHMultilocusFst(const PolymorphismMultiGContainer& pmgc, std::vector<size_t> locus_positions, const std::set<size_t>& groups);
+  static double getRHMultilocusFst(
+      const PolymorphismMultiGContainer& pmgc,
+      std::vector<size_t> locusPositions, 
+      const std::set<size_t>& groups);
 
   /**
    * @brief Compute pairwise distances on a set of groups for a given set of loci.
    * distance is either Nei72, Nei78, Fst W&C or Fst Robertson & Hill, Nm,
    * D=-ln(1-Fst) of Reynolds et al. 1983, Rousset 1997 Fst/(1-Fst)
    */
-  static std::unique_ptr<DistanceMatrix> getDistanceMatrix(const PolymorphismMultiGContainer& pmgc, std::vector<size_t> locus_positions, const std::set<size_t>& groups, std::string distance_method);
+  static std::unique_ptr<DistanceMatrix> getDistanceMatrix(
+      const PolymorphismMultiGContainer& pmgc, 
+      std::vector<size_t> locusPositions,
+      const std::set<size_t>& groups, 
+      std::string distance_method);
 };
 } // end of namespace bpp;
 
